Revert "Sprint 29 Release Changes (#32)" (#33)
This reverts commit da7526af736c53b954f67aaa1ed5e078d327210f.

+++ engineering/models.py
@@ -119,7 +119,7 @@ class Cycle(models.Model):
             return f"Cycle ({self.start_date} to {self.end_date})"
 
 class EngineerPoint(TimeStampedModel):
-    points = models.DecimalField(_('points'), default=0.0, max_digits=10 ,decimal_places=2)
+    points = models.IntegerField(_('points'), default=0)
     is_active = models.BooleanField(_('Is active'), default=True)
     cycle = models.ForeignKey(
         Cycle, on_delete=models.PROTECT,
